+++ cursedspace/__init__.py
@@ -4,6 +4,7 @@ from .key import Key
 from .panel import Panel
 from .scrollpanel import ScrollPanel
 from .inputline import InputLine, Completion
+from .inputline import SingleWordCompletion, SingleWordCompletionWrapper
 from .shellcontext import ShellContext
 from .grid import Grid
 from .progress import ProgressBar